+++ src/tb_watcher/driver_utils.py
@@ -9,23 +9,25 @@ import time
 import json
 import random
 from abc import abstractmethod
-from urllib.parse import urlparse
 
-from typing import Callable, List
+from typing import Callable, List, Union
 from dataclasses import dataclass, asdict
 
 # tb_watcher
 from tb_watcher.logger import logger
+from tb_watcher.threading import add_job
 
 # selenium
 import selenium
 from selenium import webdriver
 from selenium.webdriver.chrome.service import Service
 from webdriver_manager.chrome import ChromeDriverManager
-from selenium.webdriver.support.ui import WebDriverWait
-from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.common.by import By
 from selenium.webdriver.common.keys import Keys
+from selenium.webdriver.support import expected_conditions as EC
+from selenium.webdriver.support.ui import WebDriverWait
+
+DEF_WINDOW_SIZE = (800, 1440)
 
 def ensures_or(f: str, otherwise: str = "NULL"):
     try:
@@ -56,6 +58,7 @@ class Tweet(Unique):
     like_count: str
     reply_count: str
     potential_boost: bool
+    parent_id: Union[str, None]
 
     def get_url(self):
         return "https://www.twitter.com/{clean_handle}/status/{id}".format(self.handle[1:], self.id)
@@ -84,7 +87,7 @@ class MaxCapturesReached(RuntimeError):
 
 def tweet_dom_get_basic_metadata(tweet_dom):
     """Retrieves all metadata from tweet dom except unique id."""
-    tm = {"id": "null"}
+    tm = {"id": "null", "parent_id": None}
     tm["tag_text"] = ensures_or(lambda: tweet_dom.find_element(By.CSS_SELECTOR,'div[data-testid="User-Names"]').text)
     try:
         splts = str(tm["tag_text"]).split("\n")
@@ -119,7 +122,10 @@ class TweetExtractor:
         self.boosted_tracker = set() # Used for tracking boosted tweets by saving tweet texts.
         self.recommended_tweets_height = None
 
+        # Used for chaining tweets as threads.
+        self.prev_tweet = None
         self.tweets_tracker = set()
+        self.tweets_ordered = []
         self.root_dir = root_dir
         self.max_captures = max_captures
 
@@ -167,7 +173,7 @@ class TweetExtractor:
             driver.switch_to.window(new_window)
 
             # Clicking on a tweet guarantees it to be a TwitterThread page.
-            tt = TwitterThread(current_tweet_data, self.root_dir, driver.current_url, fetch_threads=fetch_threads, existing_driver=driver)
+            tt = TwitterThread(current_tweet_data, self.prev_tweet, self.root_dir, driver.current_url, fetch_threads=fetch_threads, existing_driver=driver)
             tm = tt.fetch_metadata()
             if tm.tweet_text != "NULL":
                 if tm.tweet_text in self.boosted_tracker:
@@ -183,14 +189,43 @@ class TweetExtractor:
             else:
                 tm.potential_boost = False
 
+            self.prev_tweet = tm
             if fetch_threads > 0:
                 logger.debug("Thread depth {}".format(fetch_threads))
-                # TODO, save to file
-                threads = tt.fetch_tweets(
-                    self.max_captures,
-                    load_time,
-                    offset_func
-                )
+                # Spawn a new thread and leave it.
+                current_url = str(driver.current_url)
+                def _new_thread(is_new_thread: bool):
+                    if is_new_thread:
+                        new_driver = create_chrome_driver()
+                        new_driver.get(current_url)
+                        # Wait for driver oload new page
+                        state = ""
+                        while state != "complete":
+                            time.sleep(random.uniform(3, 5))
+                            state = new_driver.execute_script("return document.readyState")
+
+                        WebDriverWait(new_driver, 30).until(EC.presence_of_element_located(
+                            (By.CSS_SELECTOR, '[data-testid="tweet"]')))
+                    else:
+                        new_driver = driver
+
+                    try:
+                        tt = TwitterThread(
+                            current_tweet_data,
+                            self.prev_tweet,
+                            self.root_dir,
+                            current_url,
+                            fetch_threads=fetch_threads,
+                            existing_driver=new_driver)
+                        tt.fetch_tweets(
+                            self.max_captures,
+                            load_time,
+                            offset_func
+                        )
+                    finally:
+                        if is_new_thread:
+                            new_driver.close()
+                add_job(_new_thread)
             else:
                 logger.debug("Thread depth reached.")
 
@@ -284,6 +319,7 @@ class TweetExtractor:
                 # Create a tweet's folder
                 self.counter += 1
                 self.tweets_tracker.add(full_dtm)
+                self.tweets_ordered.append(full_dtm)
 
                 if self.max_captures and self.counter > self.max_captures:
                     exit_loop = True
@@ -291,7 +327,7 @@ class TweetExtractor:
 
     def get_tweets_as_dict(self) -> List[dict]:
         results = []
-        for t in self.tweets_tracker:
+        for t in self.tweets_ordered:
             results.append(asdict(t))
         return results
 
@@ -400,4 +436,6 @@ def create_chrome_driver() -> webdriver:
     """Creates a chrome driver with silenced warnings and custom options."""
     options = webdriver.ChromeOptions()
     options.add_experimental_option('excludeSwitches', ['enable-logging'])
-    return webdriver.Chrome(options=options, service=Service(ChromeDriverManager().install()))
+    driver = webdriver.Chrome(options=options, service=Service(ChromeDriverManager().install()))
+    driver.set_window_size(*DEF_WINDOW_SIZE)
+    return driver