+++ CHANGELOG.md
@@ -18,9 +18,12 @@
 
 # Changes
 
-## 0.6.0: Twitter Thread Support
+## 0.6.0: Threaded Update
 
 * Added `--depth` and `-d` for archiving threads.
+* Added `--multi-threading` and `-t` to specify number of threads to use.
+* Added `parent_id` to `tweets.json` which tracks direct tweet responses.
+* Added `threading.py` which manages thread workers and queues.
 * Added `pages.py` which abstracts away a page on Twitter. Useful for multi-threading.
     * Added `TwitterPage` used for representing a page in Twitter.
     * Added `TwitterBio` which is used for fetching and obtaining bio pages.